mod fctmat

+++ src/potagibus.h
@@ -14,7 +14,7 @@ bool excl_carre(coord* point, coord* dep, coord* arr, int marge);
 
 bool excl_ovale(coord* point, coord* dep, coord* arr, int marge);
 
-int** create_Matrice(int n);
+float** create_Matrice(int n);
 
 float distance(coord* p1, coord* p2);
